+++ src/module.ts
@@ -1,19 +1,19 @@
-import { NgModule, ModuleWithProviders } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { IntercomBrowser } from './providers/intercom.browser';
-import { Intercom, IntercomEnvironment } from './providers/intercom';
+import { ModuleWithProviders, NgModule } from '@angular/core';
+import { Intercom } from './providers/intercom';
+import { IntercomConfig } from './types/intercom-config';
+import loadScripts from './load-intercom.js';
 
 
 @NgModule({
     imports: [CommonModule]
 })
 export class IntercomModule {
-    static forRoot(): ModuleWithProviders {
+    static forRoot(config: IntercomConfig): ModuleWithProviders {
+        loadScripts(config.app_id);
         return {
             ngModule: IntercomModule,
             providers: [
-                IntercomBrowser,
-                { provide: IntercomEnvironment, useClass: IntercomBrowser },
                 Intercom
             ]
         }